+++ CHANGELOG.md
@@ -13,7 +13,7 @@ the [releases page](https://github.com/Consensys/teku/releases).
 
 ### Additions and Improvements
 - Add `proposer_slashing`, `attester_slasing` and `payload_attibutes` (only when a loaded validator will be producing a block) events to the `/eth/v1/events` SSE stream
-- Added `--stop-vc-when-validator-slashed` option to stop the Validator Client when any owned validator is slashed (requires a BN that publishes the `attester_slashing` and `proposer_slashing` SSE events)
+- Added `--stop-vc-when-validator-slashed-enabled` option to stop the Validator Client when any owned validator is slashed (requires a BN that publishes the `attester_slashing` and `proposer_slashing` SSE events)
 
 ### Bug Fixes
 - Fix incompatibility between Teku validator client and Lighthouse beacon nodes